Extracting resolving into a separate class
